Sprite bug fixed!
Huge thanks to Rangi and all her help. The problem is with the order in which pokecrystal loads sprites by default. For future reference, reordering object_events so that the custom sprite is on the bottom fixed it completely.

A more permanent fix will be available later, it has something to do with LoadAndSortSprites